Pool Man fully reworked (see descr.)
PoolMan now only spawns 5 of each base unit (and each variant), to allow quick store refresh. When a unit is bought it will need to be instantiated at that point as well. This decreases number of entities by up to 9 times. Also these "Store Units" are now frozen on the store once they exist on all Clients